pyproject.toml
[tool.poetry]
name = "dna-frame-translation-plugin"
version = "1.0.0"
description = "DNA/RNA Frame Translation Plugin for Cauldron"
authors = ["Toan Phung"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
click = "^8.0.0"
biopython = "^1.81"
pandas = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
